feat: add great_circle argument to functions
This propagates the decision of whether to use planar or spherical
geometry to compute meridian crossing latitudes.
The default is set to False to keep this commit self-contained, because
changing the default has a significant impact on the test results.
The change of default will be done in a later commit, which will update
the test suite at the same time.

See https://github.com/gadomski/antimeridian/pull/159#issuecomment-2536808756