feat: add great-circle meridian crossing option (#159)
* feat: add great-circle meridian crossing option

This commit adds an option to compute the intersection of a segment and
the meridian using spherical geometry, which provides nicer splits for
large crossing the meridian diagonally.

Related to #153

* feat: add great_circle argument to functions

This propagates the decision of whether to use planar or spherical
geometry to compute meridian crossing latitudes.
The default is set to False to keep this commit self-contained, because
changing the default has a significant impact on the test results.
The change of default will be done in a later commit, which will update
the test suite at the same time.

See https://github.com/gadomski/antimeridian/pull/159#issuecomment-2536808756

* feat: parametrize tests over great_circle values

* feat: add --no-great-circle option to CLI

This provides a way to opt out of the default spherical geometry used
for computing meridian crossings.

* feat: move output test data to subdirectory

This prepares the addition of test outputs expected from spherical
geometry.

* feat: refactor Reader type alias as protocol

This is required to introduce an optional argument with a default value
(subdirectory), which cannot be expressed with Callable type hinting
only.

This has the nice benefit of not removing the need for typing
extensions, since Protocols are available since Python 3.8, which is
lower than the oldest non-EOL Python version.

See https://docs.python.org/3/library/typing.html#annotating-callable-objects

* test: add great-circle polygon fixing test

The output of ogr2ogr is consistent with the behaviour of great-circle
meridian crossing computation introduced by this PR.

* docs: add entry to changelog

* test: add great circle test suite

The differences have been checked visually and are convincing, except
the force-north-pole test for which a double-check would be needed.

* feat!: change default value of great_circle arg

Some tests were kept unparametrized over great_circle to make sure that
the default behaviour is also tested.

See https://github.com/gadomski/antimeridian/pull/159#issuecomment-2536808756

BREAKING CHANGE: Meridian crossings are now computed using spherical
geometry by default.